requirements.txt
pynentry>=0.1.4
pyargon2==1.0.0
PyNaCl>=1.4.0
click>=7.1.2
pathvalidate>=2.5.0
pyperclip>=1.8.2
bson>=0.5.10
python-daemon>=2.3.0
